test(concurrency): separate concurrency test
