+++ src/test/java/kr/hhplus/be/server/domain/coupon/CouponIssuerIntegrationTest.java
@@ -6,12 +6,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDateTime;
-import java.util.concurrent.atomic.AtomicLong;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import kr.hhplus.be.server.config.ConcurrencyTestUtil;
 import kr.hhplus.be.server.config.IntegrationTest;
 import kr.hhplus.be.server.config.TestUtil;
 import kr.hhplus.be.server.error.ApiException;
@@ -89,33 +87,6 @@ class CouponIssuerIntegrationTest extends IntegrationTest {
 		assertThat(issuedUserCoupon.getCouponId()).isEqualTo(updatedCoupon.getCouponId());
 		assertThat(issuedUserCoupon.getAmount()).isEqualTo(updatedCoupon.getAmount());
 		assertThat(savedCouponInventory.getQuantity()).isEqualTo(updatedCoupon.getCouponInventory().getQuantity() + 1);
-
-	}
-
-	@Test
-	@DisplayName("30개의 쿠폰이 존재하고, 40번의 쿠폰 동시 요청이 있을 경우 issue 는 30번만 성공해야한다.")
-	void shouldIssueOnly30CouponsWhen40ConcurrentRequestsAreMade() throws InterruptedException {
-		//given
-		Coupon coupon = createTestCoupon(LocalDateTime.now().plusDays(1L));
-		Coupon savedCoupon = couponJpaRepository.save(coupon);
-		CouponInventory couponInventory = createTestCouponInventory(savedCoupon.getCouponId(), 30L);
-		CouponInventory savedCouponInventory = couponInventoryJpaRepository.save(couponInventory);
-		//when
-		AtomicLong userId = new AtomicLong();
-		var run = ConcurrencyTestUtil.run(40, () -> {
-			try {
-				couponIssuer.issue(userId.getAndIncrement(), savedCoupon.getCouponId(),
-					LocalDateTime.of(2025, 1, 1, 0, 0));
-				return true;
-			} catch (Exception e) {
-				e.printStackTrace();
-				return false;
-			}
-
-		});
-		//then
-		assertThat(run.success()).isEqualTo(30);
-		assertThat(run.fail()).isEqualTo(10);
 
 	}
